pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-analyser"
version = "0.1.0"
description = "Semantic code analysis system that understands code by meaning, not just keywords"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Sameer"}
]
dependencies = [
    "chromadb>=0.4.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tree-sitter>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
]

[project.scripts]
code-analyser = "src.cli.main:cli"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
